Changed init of result vectors

+++ src/BLAS_MC/Lin_Functions/SYMV.jl
@@ -14,8 +14,8 @@ function SYMV(UPLO::String, n::Integer, alpha::Float64,  A::Array{MC{N},2}, x::A
 #Skip testing parameters
 MCzero::MC = MC{N}(0.0, 0.0)
 #Not using sparse vectors, assume incx,incy==1
-kx =1
-ky =1
+kx::Int =1
+ky::Int =1
 #Form beta*y
 y_2::Array{MC{N},1} = fill(MCzero, length(y)) #Result vector. dont want to solve in place of right now
 if beta != 1#If 1 can ignore coefficient
